% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_supt_bounds}
\alias{calculate_supt_bounds}
\title{Calculate sup-t bounds}
\usage{
calculate_supt_bounds(estimates, var, alpha)
}
\arguments{
\item{estimates}{A vector of point estimates}

\item{var}{The variance-covariance matrix of the estimates}

\item{alpha}{Significance level}
}
\value{
A list containing lower bounds, upper bounds, and critical value
}
\description{
Calculate sup-t bounds
}
\keyword{internal}
