% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_bounds.R
\name{print.plausible_bounds}
\alias{print.plausible_bounds}
\title{Print method for plausible_bounds objects}
\usage{
\method{print}{plausible_bounds}(x, ...)
}
\arguments{
\item{x}{A plausible_bounds object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for plausible_bounds objects
}
