% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_bounds.R
\name{print.summary.plausible_bounds}
\alias{print.summary.plausible_bounds}
\title{Print method for summary.plausible_bounds objects}
\usage{
\method{print}{summary.plausible_bounds}(x, ...)
}
\arguments{
\item{x}{A summary.plausible_bounds object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input summary.plausible_bounds object and displays summary data.frame of restricted estimates and plausible bounds.
}
\description{
Print method for summary.plausible_bounds objects
}
\examples{
# Example with constant design and no correlation
data(estimates_constant)
data(var_constant)
pb <- plausible_bounds(estimates_constant[1:4], var_constant[1:4, 1:4])
pb_df <- summary(pb)
print(pb_df)

}
