% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_constant}
\alias{var_constant}
\title{Variance Matrix for Constant Estimates}
\format{
A 12 x 12 matrix
}
\source{
Generated from simulation with constant design and no correlation across horizons.
}
\usage{
var_constant
}
\description{
A variance matrix for the constant estimates.
This is a diagonal matrix representing no correlation across time periods.
}
\examples{
data(estimates_constant)
data(var_constant)
result <- plausible_bounds(estimates_constant[1:4], var_constant[1:4, 1:4])
create_plot(result)
}
\keyword{datasets}
