% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimates_constant}
\alias{estimates_constant}
\title{Constant Estimates}
\format{
A numeric vector with 12 elements
}
\source{
Generated from simulation with constant design and no correlation across horizons
}
\usage{
estimates_constant
}
\description{
A dataset containing estimates from a simple constant design, no correlation across horizons.
}
\examples{
data(estimates_constant)
data(var_constant)
result <- plausible_bounds(estimates_constant[1:4], var_constant[1:4, 1:4])
create_plot(result)
}
\keyword{datasets}
