% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimates_smooth}
\alias{estimates_smooth}
\title{Smooth Estimates from Freyaldenhoven and Hansen (2026)}
\format{
A numeric vector with 44 elements (8 preperiods, 36 postperiods)
}
\source{
Point estimates from Figure 1 of Freyaldenhoven and Hansen (2026)
}
\usage{
estimates_smooth
}
\description{
A dataset containing smooth treatment effect estimates that dip down and then converge to zero.
The first 8 observations comprise the preperiods, the next 36 are post-period.
}
\examples{
data(estimates_smooth)
data(var_smooth)
result <- plausible_bounds(estimates_smooth[9:13], var_smooth[9:13, 9:13])
create_plot(result)
}
\keyword{datasets}
