% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_bounds.R
\name{print.plausible_bounds}
\alias{print.plausible_bounds}
\title{Print method for plausible_bounds objects}
\usage{
\method{print}{plausible_bounds}(x, ...)
}
\arguments{
\item{x}{A plausible_bounds object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns a plausible_bounds object and displays summary of main results
}
\description{
Print method for plausible_bounds objects
}
\examples{
# Example with constant estimates and no correlation (simple case)
data(estimates_constant)
data(var_constant)
pb <- plausible_bounds(estimates_constant[1:4], var_constant[1:4, 1:4])
print(pb)

}
