% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_bounds.R
\name{print.summary.plausible_bounds}
\alias{print.summary.plausible_bounds}
\title{Print method for summary.plausible_bounds objects}
\usage{
\method{print}{summary.plausible_bounds}(x, ...)
}
\arguments{
\item{x}{A summary.plausible_bounds object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for summary.plausible_bounds objects
}
