% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_bounds.R
\name{summary.plausible_bounds}
\alias{summary.plausible_bounds}
\title{Summary method for plausible_bounds objects}
\usage{
\method{summary}{plausible_bounds}(object, ...)
}
\arguments{
\item{object}{A plausible_bounds object}

\item{...}{Additional arguments passed to summary}
}
\description{
Summary method for plausible_bounds objects
}
