% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_iid}
\alias{var_iid}
\title{Variance Matrix for Constant Estimates with IID Errors}
\format{
A 12 x 12 matrix
}
\source{
Generated from simulation with constant design and IID errors
}
\usage{
var_iid
}
\description{
A variance matrix for the constant, independent, and identically distributed estimates.
This is a diagonal matrix representing no correlation across time periods.
}
\examples{
data(estimates_constant)
data(var_iid)
result <- plausible_bounds(estimates_constant[1:4], var_iid[1:4, 1:4])
create_plot(result)
}
\keyword{datasets}
