% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_smooth}
\alias{var_smooth}
\title{Variance Matrix for Smooth Estimates}
\format{
A 44 x 44 matrix
}
\source{
Variance Matrix of Figure 1 of Freyaldenhoven and Hansen (2026)
}
\usage{
var_smooth
}
\description{
A variance matrix for the smooth estimates. This matrix captures
the correlation structure of the estimates.
The first 8 rows and 8 columns comprise the variance matrix for the 8 preperiods.
}
\examples{
data(estimates_smooth)
data(var_smooth)
result <- plausible_bounds(estimates_smooth[9:13], var_smooth[9:13, 9:13])
create_plot(result)
}
\keyword{datasets}
